refactor: centralize config generation
